This  is a vast project containing number of java and the text file required for the operation, It also contains a folder used for tracking each and every sale of the food item respective to its category.
There are three main classes in the project and they go as RunnerPizzerio.java, PizzerioVegPizza.java and PizzasOrder.java all the files are well maintained and are encapsulated in which RunnerPijzzerio.java acts as the main java file for accepting inputs, giving the output and transfering the logic to PizzasOrder.java which is the brain of the project and contains all the main and important executions, then comes the PizzerioVegPizzas.java which handles the calculations of the bill and is the most safest class  of the project.
Other are the text file where we can add or remove the food items the same changes will be observed in the code by its own and other text file are used for tacking the count of the sale of each foood items each time.